Restructuring for easier testing.
