+++ lib/tdtool.js
