+++ spec/telldus-platform.spec.js
@@ -24,7 +24,7 @@ describe('test', () => {
   it('should create a service when accessed by HomeBridge', () => {
     registerInjector({
       'tellstick-confparser': jasmine.createSpy(),
-      './td-tool': {
+      './lib/tdtool': {
         listDevices: jasmine.createSpy().and.returnValue(null)
       }
     })
@@ -35,7 +35,7 @@ describe('test', () => {
     it('should log no found items.', done => {
       registerInjector({
         'tellstick-confparser': jasmine.createSpy(),
-        './td-tool': {
+        './lib/tdtool': {
           listDevices: jasmine.createSpy().and.returnValue(Promise.resolve([]))
         }
       })
@@ -44,7 +44,31 @@ describe('test', () => {
         expect(accessories).toEqual([])
         expect(log.calls.allArgs()[0][0]).toEqual('Loading devices...')
         expect(log.calls.allArgs()[1][0]).toEqual(
-          'Found 0 items of type "device".')
+          'Found no items of type "device".')
+        done()
+      })
+    })
+
+    it('should only create accessories that are of type "device".', done => {
+      const devices = [
+        {type: 'device'},
+        {type: 'not-a-device'},
+      ]
+
+      registerInjector({
+        'tellstick-confparser': jasmine.createSpy(),
+        './lib/tdtool': {
+          listDevices: jasmine.createSpy().and.returnValue(
+            Promise.resolve(devices)),
+          device: jasmine.createSpy().and.callFake(id =>
+            Promise.resolve(devices.find(d => d.id === id)))
+        }
+      })
+
+      instance.accessories(accessories => {
+        expect(log.calls.allArgs()[0][0]).toEqual('Loading devices...')
+        expect(log.calls.allArgs()[1][0]).toEqual(
+          'Found 1 item of type "device".')
         done()
       })
     })
@@ -52,7 +76,7 @@ describe('test', () => {
     it('should create accessories when passed.', done => {
       const devices = [{
         id: 1,
-        name: 'Livingroom:lamps-window',
+        name: 'Lamp1',
         controller: 0,
         protocol: 'arctech',
         model: 'selflearning-switch:nexa',
@@ -60,7 +84,7 @@ describe('test', () => {
         lastsentcommand: 'ON',
       }, {
         id: 2,
-        name: 'Livingroom:lamps-cabinet',
+        name: 'Lamp2',
         controller: 0,
         protocol: 'arctech',
         model: 'selflearning-switch:nexa',
@@ -70,7 +94,7 @@ describe('test', () => {
 
       registerInjector({
         'tellstick-confparser': jasmine.createSpy(),
-        './td-tool': {
+        './lib/tdtool': {
           listDevices: jasmine.createSpy().and.returnValue(
             Promise.resolve(devices)),
           device: jasmine.createSpy().and.callFake(id =>
